python/requirements.txt
paho-mqtt
simpleaudio
PyAudio
